Students controller changes

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -16,13 +16,11 @@ namespace ContosoUniversity.Controllers
 			_context = context;
 		}
 
-		// GET: Students
 		public async Task<IActionResult> Index()
 		{
 			return View(await _context.Students.ToListAsync());
 		}
 
-		// GET: Students/Details/5
 		public async Task<IActionResult> Details(int? id)
 		{
 			if (id == null)
@@ -31,7 +29,11 @@ namespace ContosoUniversity.Controllers
 			}
 
 			var student = await _context.Students
-				.FirstOrDefaultAsync(m => m.Id == id);
+				.Include(s => s.Enrollments)
+				.ThenInclude(e => e.Course)
+				.AsNoTracking()
+				.SingleOrDefaultAsync(m => m.Id == id);
+
 			if (student == null)
 			{
 				return NotFound();
@@ -40,7 +42,6 @@ namespace ContosoUniversity.Controllers
 			return View(student);
 		}
 
-		// GET: Students/Create
 		public IActionResult Create()
 		{
 			return View();
@@ -51,18 +52,27 @@ namespace ContosoUniversity.Controllers
 		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> Create([Bind("Id,LastName,FirstMidName,EnrollmentDate")] Student student)
+		public async Task<IActionResult> Create([Bind("LastName,FirstMidName,EnrollmentDate")] Student student)
 		{
-			if (ModelState.IsValid)
+			try
 			{
-				_context.Add(student);
-				await _context.SaveChangesAsync();
-				return RedirectToAction(nameof(Index));
+				if (ModelState.IsValid)
+				{
+					_context.Add(student);
+					await _context.SaveChangesAsync();
+					return RedirectToAction(nameof(Index));
+				}
+			}
+			catch (DbUpdateException /* ex */)
+			{
+				//Log the error (uncomment ex variable name and write a log.
+				ModelState.AddModelError("", "Unable to save changes. " +
+				                             "Try again, and if the problem persists " +
+				                             "see your system administrator.");
 			}
 			return View(student);
 		}
 
-		// GET: Students/Edit/5
 		public async Task<IActionResult> Edit(int? id)
 		{
 			if (id == null)
@@ -81,38 +91,37 @@ namespace ContosoUniversity.Controllers
 		// POST: Students/Edit/5
 		// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
 		// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-		[HttpPost]
+		[HttpPost, ActionName("Edit")]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> Edit(int id, [Bind("Id,LastName,FirstMidName,EnrollmentDate")] Student student)
+		public async Task<IActionResult> EditPost(int? id)
 		{
-			if (id != student.Id)
+			if (id == null)
 			{
 				return NotFound();
 			}
-
-			if (ModelState.IsValid)
+			var studentToUpdate = await _context.Students.SingleOrDefaultAsync(s => s.Id == id);
+			if (await TryUpdateModelAsync(
+				studentToUpdate,
+				"",
+				s => s.FirstMidName, s => s.LastName, s => s.EnrollmentDate))
 			{
 				try
 				{
-					_context.Update(student);
 					await _context.SaveChangesAsync();
+					return RedirectToAction(nameof(Index));
 				}
-				catch (DbUpdateConcurrencyException)
+				catch (DbUpdateException /* ex */)
 				{
-					if (!StudentExists(student.Id))
-					{
-						return NotFound();
-					}
-
-					throw;
+					//Log the error (uncomment ex variable name and write a log.)
+					ModelState.AddModelError("", "Unable to save changes. " +
+					                             "Try again, and if the problem persists, " +
+					                             "see your system administrator.");
 				}
-				return RedirectToAction(nameof(Index));
 			}
-			return View(student);
+			return View(studentToUpdate);
 		}
 
-		// GET: Students/Delete/5
-		public async Task<IActionResult> Delete(int? id)
+		public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
 		{
 			if (id == null)
 			{
@@ -120,24 +129,46 @@ namespace ContosoUniversity.Controllers
 			}
 
 			var student = await _context.Students
-				.FirstOrDefaultAsync(m => m.Id == id);
+				.AsNoTracking()
+				.SingleOrDefaultAsync(m => m.Id == id);
 			if (student == null)
 			{
 				return NotFound();
 			}
 
+			if (saveChangesError.GetValueOrDefault())
+			{
+				ViewData["ErrorMessage"] =
+					"Delete failed. Try again, and if the problem persists " +
+					"see your system administrator.";
+			}
+
 			return View(student);
 		}
 
-		// POST: Students/Delete/5
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> DeleteConfirmed(int id)
 		{
-			var student = await _context.Students.FindAsync(id);
-			_context.Students.Remove(student);
-			await _context.SaveChangesAsync();
-			return RedirectToAction(nameof(Index));
+			var student = await _context.Students
+				.AsNoTracking()
+				.SingleOrDefaultAsync(m => m.Id == id);
+			if (student == null)
+			{
+				return RedirectToAction(nameof(Index));
+			}
+
+			try
+			{
+				_context.Students.Remove(student);
+				await _context.SaveChangesAsync();
+				return RedirectToAction(nameof(Index));
+			}
+			catch (DbUpdateException /* ex */)
+			{
+				//Log the error (uncomment ex variable name and write a log.)
+				return RedirectToAction(nameof(Delete), new {id, saveChangesError = true });
+			}
 		}
 
 		private bool StudentExists(int id)
